Move utils code elsewhere Move demo code to examples folder Add bash script to download demo data Start of Readme
