+++ trl/trainer/soft_q_learning_trainer.py
@@ -8,6 +8,7 @@ import torch
 import torch.nn.functional as F
 
 from ..models import PreTrainedModelWrapper
+from ..core import mask_and_reduce, masked_reverse_cumsum, get_masked_mean_min_max, entropy_from_logits
 from . import BaseTrainer
 
 
@@ -44,7 +45,9 @@ class SoftQLearningTrainer(BaseTrainer):
         top_k: Optional[int] = None,
         top_p: Optional[float] = None,
         beam_width: Optional[int] = None,
-        reward_function: Callable[[List[str], List[str], List[str]], Tuple[FloatTensor, Dict[str, Any]]] = None,
+        reward_function: Callable[
+            [List[str], List[str], List[str]], Tuple[FloatTensor, Dict[str, Any]]
+        ] = None,
         tokenizer=None,
         device="cuda:0",
         reward_shaping_func: Callable[[FloatTensor], FloatTensor] = lambda r: r,
@@ -78,9 +81,11 @@ class SoftQLearningTrainer(BaseTrainer):
 
         trainable_params = [p for p in model.parameters() if p.requires_grad]
         if not target_model_empty:
-            trainable_params.extend([p for p in self.target_model.parameters() if p.requires_grad])
+            trainable_params.extend(
+                [p for p in self.target_model.parameters() if p.requires_grad]
+            )
 
-        self.optimizer = torch.optim.Adam(trainable_params, lr=1e-4)
+        self.optimizer = torch.optim.Adam(trainable_params, lr=target_learning_rate)
 
     def step(self, batch, step):
         # if PREPROCESS_TARGET_TEXTS is True:
@@ -145,20 +150,27 @@ class SoftQLearningTrainer(BaseTrainer):
         # would yield the same parameter orders.
         # https://towardsdatascience.com/double-deep-q-networks-905dd8325412
         elif sync_type == "polyak":
-            for param_, param in zip(self.target_model.parameters(), self.model.parameters()):
-
-                param_.data.copy_((1 - self.target_learning_rate) * param_ + self.target_learning_rate * param)
+            for param_, param in zip(
+                self.target_model.parameters(), self.model.parameters()
+            ):
+                param_.data.copy_(
+                    (1 - self.target_learning_rate) * param_
+                    + self.target_learning_rate * param
+                )
         else:
             # log warning
             # TODO: come back here
             return
 
     def _forward_SQL(self, mode, batch):
-
         if mode == ForwardMode.SQL_OFF:
             # teacher forcing
-            outputs = self.model(input_ids=batch[0].input_ids, decoder_input_ids=batch[1].input_ids)
-            target_outputs = self.target_model(input_ids=batch[0].input_ids, decoder_input_ids=batch[1].input_ids)
+            outputs = self.model(
+                input_ids=batch[0].input_ids, decoder_input_ids=batch[1].input_ids
+            )
+            target_outputs = self.target_model(
+                input_ids=batch[0].input_ids, decoder_input_ids=batch[1].input_ids
+            )
 
             logits = outputs.logits.to(self.device)
             target_logits = target_outputs.logits.to(self.device)
@@ -166,7 +178,6 @@ class SoftQLearningTrainer(BaseTrainer):
             sequence_lengths = batch[1].attention_mask.sum(dim=-1).to(self.device)
 
         elif mode == ForwardMode.SQL_ON:
-
             generation_length = batch[1].input_ids.shape[1]
 
             outputs = self.model.generate(
@@ -184,23 +195,34 @@ class SoftQLearningTrainer(BaseTrainer):
 
             # has to follow the steps taken by the model
             target_outputs = self.target_model(
-                input_ids=on_policy_training_data[0], decoder_input_ids=on_policy_training_data[1]
+                input_ids=on_policy_training_data[0],
+                decoder_input_ids=on_policy_training_data[1],
             )
 
             # possible problematic spot
             logits = torch.stack(outputs.scores, dim=1).to(self.device)
             target_logits = target_outputs.logits.to(self.device)
             output_ids = outputs.sequences[:, 1:].contiguous().to(self.device)
-            sequence_lengths = (output_ids != self.tokenizer.pad_token_id).sum(dim=-1).to(self.device)
+            sequence_lengths = (
+                (output_ids != self.tokenizer.pad_token_id).sum(dim=-1).to(self.device)
+            )
 
         else:
             raise NotImplementedError
 
-        predicted_texts = self.tokenizer.batch_decode(output_ids, skip_special_tokens=True)
-        input_texts = self.tokenizer.batch_decode(batch[0].input_ids, skip_special_tokens=True)
-        target_texts = self.tokenizer.batch_decode(batch[1].input_ids, skip_special_tokens=True)
+        predicted_texts = self.tokenizer.batch_decode(
+            output_ids, skip_special_tokens=True
+        )
+        input_texts = self.tokenizer.batch_decode(
+            batch[0].input_ids, skip_special_tokens=True
+        )
+        target_texts = self.tokenizer.batch_decode(
+            batch[1].input_ids, skip_special_tokens=True
+        )
 
-        raw_rewards, shaped_rewards, rewards_log = self.compute_rewards(input_texts, target_texts, predicted_texts)
+        raw_rewards, shaped_rewards, rewards_log = self.compute_rewards(
+            input_texts, target_texts, predicted_texts
+        )
 
         sql_loss, sql_loss_log = soft_q_loss_with_sparse_rewards(
             implementation=self.sql_loss_impl,
@@ -212,8 +234,14 @@ class SoftQLearningTrainer(BaseTrainer):
             coefficient=self.sql_loss_coefficients,
             # Do not add margin losses unless the
             # actions are ground truth actions.
-            margin_constant=(self.sql_loss_margin_constant if mode == ForwardMode.SQL_OFF else None),
-            margin_coefficient=(self.sql_loss_margin_coefficient if mode == ForwardMode.SQL_OFF else None),
+            margin_constant=(
+                self.sql_loss_margin_constant if mode == ForwardMode.SQL_OFF else None
+            ),
+            margin_coefficient=(
+                self.sql_loss_margin_coefficient
+                if mode == ForwardMode.SQL_OFF
+                else None
+            ),
         )
 
         add_prefix_to_dict_keys_inplace(rewards_log, prefix=f"{mode.value}/rewards/")
@@ -237,7 +265,11 @@ class SoftQLearningTrainer(BaseTrainer):
 
     def compute_rewards(self, source_texts, target_texts, output_texts):
         rewards_tensor, rewards_log = self.reward_function(
-            sources=source_texts, targets=target_texts, predictions=output_texts, to_tensor=True, mode="train"
+            sources=source_texts,
+            targets=target_texts,
+            predictions=output_texts,
+            to_tensor=True,
+            mode="train",
         )
 
         rewards_tensor = rewards_tensor  # .to(self.device)
@@ -248,29 +280,6 @@ class SoftQLearningTrainer(BaseTrainer):
         raise NotImplementedError("Not implemented")
 
 
-# ====================================================================================================
-
-# = Soft Q Learning Loss Functions with Sparse Rewards (TODO: note about where it is copied from )
-
-# =============================================
-
-
-# TODO: find a pytorch implementation of this
-def get_entropy(logits: torch.Tensor) -> torch.Tensor:
-    r"""Compute entropy according to the definition.
-
-    Args:
-        logits: Unscaled log probabilities.
-
-    Return:
-        A tensor containing the Shannon entropy in the last dimension.
-    """
-    probs = F.softmax(logits, -1) + 1e-8
-    entropy = -probs * torch.log(probs)
-    entropy = torch.sum(entropy, -1)
-    return entropy
-
-
 def unionize_dicts(dict_list):
     merged_dict = {}
     for dict_ in dict_list:
@@ -278,11 +287,15 @@ def unionize_dicts(dict_list):
     return merged_dict
 
 
-def gather_2d_on_last_dim(tensor: FloatTensor, index: LongTensor, shape: torch.Size) -> FloatTensor:
+def gather_2d_on_last_dim(
+    tensor: FloatTensor, index: LongTensor, shape: torch.Size
+) -> FloatTensor:
     """Simplified version of `tf.gather_nd` in PyTorch"""
     flattened_tensor = tensor.view(-1, tensor.shape[-1])
     flattened_index = index.view(-1)
-    flattened_gathered_tensor = flattened_tensor[torch.arange(flattened_index.shape[0]), flattened_index]
+    flattened_gathered_tensor = flattened_tensor[
+        torch.arange(flattened_index.shape[0]), flattened_index
+    ]
     return flattened_gathered_tensor.view(shape)
 
 
@@ -291,7 +304,6 @@ def add_prefix_to_dict_keys_inplace(
     prefix: str,
     keys_to_exclude: Optional[List[str]] = None,
 ) -> None:
-
     # https://stackoverflow.com/questions/4406501/change-the-name-of-a-key-in-dictionary
     keys = list(d.keys())
     for key in keys:
@@ -324,7 +336,15 @@ def soft_q_loss_with_sparse_rewards(
         rewards:         [batch_size]
         sequence_length: [batch_size]
     """
-    if implementation not in ["v0", "v1", "v2", "v3", "v2_v2r", "v3_v3r", "v2_v2r_v3_v3r"]:
+    if implementation not in [
+        "v0",
+        "v1",
+        "v2",
+        "v3",
+        "v2_v2r",
+        "v3_v3r",
+        "v2_v2r_v3_v3r",
+    ]:
         raise ValueError
 
     if not torch.is_tensor(rewards):
@@ -354,16 +374,28 @@ def soft_q_loss_with_sparse_rewards(
         )
 
     if implementation == "v3_v3r":
-        _sql_loss_func = partial(soft_q_loss_with_sparse_rewards_3_3_reversed, coefficient=coefficient)
+        _sql_loss_func = partial(
+            soft_q_loss_with_sparse_rewards_3_3_reversed, coefficient=coefficient
+        )
 
     if implementation == "v2_v2r_v3_v3r":
-        _sql_loss_func = partial(soft_q_loss_with_sparse_rewards_2_2_reversed_3_3_reversed, coefficient=coefficient)
+        _sql_loss_func = partial(
+            soft_q_loss_with_sparse_rewards_2_2_reversed_3_3_reversed,
+            coefficient=coefficient,
+        )
 
     if logits.shape != logits_.shape:
-        raise ValueError(f"`logits.shape` = {logits.shape}, but " f"`logits_.shape` = {logits_.shape}")
+        raise ValueError(
+            f"`logits.shape` = {logits.shape}, but "
+            f"`logits_.shape` = {logits_.shape}"
+        )
 
     raw_losses, quantities_to_log = _sql_loss_func(
-        logits=logits, logits_=logits_, actions=actions, rewards=rewards, sequence_length=sequence_length
+        logits=logits,
+        logits_=logits_,
+        actions=actions,
+        rewards=rewards,
+        sequence_length=sequence_length,
     )
 
     loss = mask_and_reduce(sequence=raw_losses, sequence_length=sequence_length)
@@ -395,7 +427,6 @@ def soft_q_loss_with_sparse_rewards_1(
     rewards: FloatTensor,
     sequence_length: LongTensor,
 ) -> Tuple[FloatTensor, Dict[str, Any]]:
-
     Q = gather_2d_on_last_dim(tensor=logits, index=actions, shape=actions.shape)
     # use `V` from the target if available
     V_ = logits_.logsumexp(dim=-1)
@@ -426,7 +457,6 @@ def soft_q_loss_with_sparse_rewards_2(
     sequence_length: LongTensor,
     _recover_mle: bool = False,
 ) -> Tuple[FloatTensor, Dict[str, Any]]:
-
     Q = gather_2d_on_last_dim(tensor=logits, index=actions, shape=actions.shape)
     V = logits.logsumexp(dim=-1)
     A = Q - V
@@ -441,7 +471,9 @@ def soft_q_loss_with_sparse_rewards_2(
     # the episode ends, thus depends on `sequence_length`
     terminal_V_ = V_[torch.arange(sequence_length.shape[0]), sequence_length - 1]
     Q_[torch.arange(sequence_length.shape[0]), sequence_length - 1] = rewards
-    A_[torch.arange(sequence_length.shape[0]), sequence_length - 1] = rewards - terminal_V_
+    A_[torch.arange(sequence_length.shape[0]), sequence_length - 1] = (
+        rewards - terminal_V_
+    )
 
     # if _recover_mle is True:
     #    sql_utils.colorful_warning("Recover-MLE Mode", bg="red")
@@ -455,8 +487,8 @@ def soft_q_loss_with_sparse_rewards_2(
         "Q_": Q_,
         "V_": V_,
         "A_": A_,
-        "H": get_entropy(logits),
-        "H_": get_entropy(logits_),
+        "H":  entropy_from_logits(logits),
+        "H_": entropy_from_logits(logits_),
     }
 
     return raw_losses, quantities_to_log
@@ -470,7 +502,6 @@ def soft_q_loss_with_sparse_rewards_3(
     sequence_length: LongTensor,
     freeze_future_steps: bool = False,
 ) -> Tuple[FloatTensor, Dict[str, Any]]:
-
     Q = gather_2d_on_last_dim(tensor=logits, index=actions, shape=actions.shape)
     V = logits.logsumexp(dim=-1)
     A = Q - V
@@ -509,16 +540,23 @@ def soft_q_loss_with_sparse_rewards_2_2_reversed(
     margin_constant: Optional[float] = None,
     margin_coefficient: Optional[float] = None,
 ) -> Tuple[FloatTensor, Dict[str, Any]]:
-
     raw_losses_2, quantities_to_log_2 = soft_q_loss_with_sparse_rewards_2(
-        logits=logits, logits_=logits_, actions=actions, rewards=rewards, sequence_length=sequence_length
+        logits=logits,
+        logits_=logits_,
+        actions=actions,
+        rewards=rewards,
+        sequence_length=sequence_length,
     )
 
     add_prefix_to_dict_keys_inplace(quantities_to_log_2, prefix="0/")
 
     if coefficient is not None:
         raw_losses_2_r, quantities_to_log_2_r = soft_q_loss_with_sparse_rewards_2(
-            logits=logits_, logits_=logits, actions=actions, rewards=rewards, sequence_length=sequence_length
+            logits=logits_,
+            logits_=logits,
+            actions=actions,
+            rewards=rewards,
+            sequence_length=sequence_length,
         )
 
         raw_losses = coefficient * raw_losses_2 + (1 - coefficient) * raw_losses_2_r
@@ -561,16 +599,23 @@ def soft_q_loss_with_sparse_rewards_3_3_reversed(
     sequence_length: LongTensor,
     coefficient: Optional[float] = None,
 ) -> Tuple[FloatTensor, Dict[str, Any]]:
-
     raw_losses_3, quantities_to_log_3 = soft_q_loss_with_sparse_rewards_3(
-        logits=logits, logits_=logits_, actions=actions, rewards=rewards, sequence_length=sequence_length
+        logits=logits,
+        logits_=logits_,
+        actions=actions,
+        rewards=rewards,
+        sequence_length=sequence_length,
     )
 
     add_prefix_to_dict_keys_inplace(quantities_to_log_3, prefix="0/")
 
     if coefficient is not None:
         raw_losses_3_r, quantities_to_log_3_r = soft_q_loss_with_sparse_rewards_3(
-            logits=logits_, logits_=logits, actions=actions, rewards=rewards, sequence_length=sequence_length
+            logits=logits_,
+            logits_=logits,
+            actions=actions,
+            rewards=rewards,
+            sequence_length=sequence_length,
         )
 
         raw_losses = coefficient * raw_losses_3 + (1 - coefficient) * raw_losses_3_r
@@ -598,7 +643,6 @@ def soft_q_loss_with_sparse_rewards_2_2_reversed_3_3_reversed(
     sequence_length: LongTensor,
     coefficient: Optional[float] = None,
 ) -> Tuple[FloatTensor, Dict[str, Any]]:
-
     raw_losses_2, quantities_to_log_2 = soft_q_loss_with_sparse_rewards_2_2_reversed(
         logits=logits,
         logits_=logits_,
@@ -660,10 +704,14 @@ def large_margin_classification_loss(
         expert_actions: [batch_size, sequence_length]
     """
     # [0, 0, 0, ..., 1, 1, 1, ..., N, N, N, ...]
-    batch_indices = torch.arange(expert_actions.shape[0]).repeat_interleave(expert_actions.shape[1], dim=0)
+    batch_indices = torch.arange(expert_actions.shape[0]).repeat_interleave(
+        expert_actions.shape[1], dim=0
+    )
 
     # [0, 1, 2, ..., 0, 1, 2, ..., 0, 1, 2, ...]
-    sequence_indices = torch.arange(expert_actions.shape[1]).repeat(expert_actions.shape[0])
+    sequence_indices = torch.arange(expert_actions.shape[1]).repeat(
+        expert_actions.shape[0]
+    )
 
     # indices for the expert actions
     indices = (batch_indices, sequence_indices, expert_actions.flatten())
@@ -673,286 +721,15 @@ def large_margin_classification_loss(
     margin[indices] = 0
 
     # [batch_size, sequence_length]
-    raw_losses = (logits + margin).max(dim=-1).values - logits[indices].view(expert_actions.shape)
+    raw_losses = (logits + margin).max(dim=-1).values - logits[indices].view(
+        expert_actions.shape
+    )
 
     quantities_to_log = {
         "loss": raw_losses,
     }
 
     return raw_losses, quantities_to_log
-
-
-def masked_reverse_cumsum(X: FloatTensor, lengths: LongTensor, dim: int) -> FloatTensor:
-    masked_X = X * sequence_mask(lengths, max_len=X.shape[1])
-
-    return masked_X.flip(dims=[dim]).cumsum(dim=dim).flip(dims=[dim])
-
-
-def get_masked_mean_min_max(X: FloatTensor, lengths: LongTensor) -> Tuple[FloatTensor, FloatTensor, FloatTensor]:
-    if X.ndim != 2 and lengths.ndim != 1:
-        raise ValueError
-
-    if X.shape[0] != lengths.shape[0]:
-        raise ValueError
-
-    mask = get_lengths_mask(X=X, lengths=lengths)
-
-    masked_min = X.masked_fill(~mask, np.inf).min(dim=1)
-    masked_max = X.masked_fill(~mask, -np.inf).max(dim=1)
-    masked_mean = mask_and_reduce(
-        sequence=X, sequence_length=lengths, average_across_timesteps=True, sum_over_timesteps=False
-    )
-
-    return (masked_mean, masked_min.values.mean(), masked_max.values.mean())
-
-
-def mask_and_reduce(
-    sequence: torch.Tensor,
-    sequence_length: torch.LongTensor,
-    rank: int = 2,
-    average_across_batch: bool = True,
-    average_across_timesteps: bool = False,
-    sum_over_batch: bool = False,
-    sum_over_timesteps: bool = True,
-    dtype: Optional[torch.dtype] = None,
-) -> torch.Tensor:
-    r"""Masks out sequence entries that are beyond the respective sequence
-    lengths, and reduces (average or sum) away dimensions.
-
-    This is a combination of :func:`~texar.torch.utils.shapes.mask_sequences`
-    and :func:`~texar.torch.losses.losses_utils.reduce_batch_time`.
-
-    Args:
-        sequence: A tensor of sequence values.
-            If `time_major=False` (default), this must be a tensor of shape
-            `[batch_size, max_time, d_2, ..., d_rank]`, where the rank of
-            the tensor is specified with :attr:`rank`.
-            The batch and time dimensions are exchanged if `time_major` is True.
-        sequence_length: A tensor of shape `[batch_size]`. Time steps beyond
-            the respective sequence lengths will be made zero. If `None`,
-            no masking is performed.
-        rank (int): The rank of :attr:`sequence`. Must be >= 2. Default is 2,
-            i.e., `sequence` is a 2D Tensor consisting of batch and time
-            dimensions.
-        average_across_timesteps (bool): If set, average the sequence across
-            the time dimension. Must not set `average_across_timesteps`
-            and `sum_over_timesteps` at the same time.
-        average_across_batch (bool): If set, average the sequence across the
-            batch dimension. Must not set `average_across_batch`'
-            and `sum_over_batch` at the same time.
-        average_across_remaining (bool): If set, average the sequence across the
-            remaining dimensions. Must not set `average_across_remaining`'
-            and `sum_over_remaining` at the same time.
-        sum_over_timesteps (bool): If set, sum the sequence across the time
-            dimension. Must not set `average_across_timesteps` and
-            `sum_over_timesteps` at the same time.
-        sum_over_batch (bool): If set, sum the sequence across the batch
-            dimension. Must not set `average_across_batch` and `sum_over_batch`
-            at the same time.
-        sum_over_remaining (bool): If set, sum the sequence across the remaining
-            dimension. Must not set `average_across_remaining` and
-            `sum_over_remaining` at the same time.
-        dtype (torch.dtype): The dtype of the returned mask.
-        time_major (bool): The shape format of the inputs. If `True`,
-            :attr:`sequence` must have shape `[max_time, batch_size, ...]`.
-            If `False` (default), `sequence` must have
-            shape `[batch_size, max_time, ...]`.
-
-    Returns:
-        A tensor containing the masked and reduced sequence.
-    """
-    if rank < 2:
-        raise ValueError("`rank` must be >= 2.")
-
-    if sequence_length is not None:
-        sequence = mask_sequences(sequence, sequence_length, dtype=dtype)
-
-    sequence = reduce_batch_time(
-        sequence, sequence_length, average_across_batch, average_across_timesteps, sum_over_batch, sum_over_timesteps
-    )
-
-    average_across_timesteps or sum_over_timesteps
-    average_across_batch or sum_over_batch
-
-    return sequence
-
-
-def reduce_batch_time(
-    sequence: torch.Tensor,
-    sequence_length: torch.LongTensor,
-    average_across_batch: bool = True,
-    average_across_timesteps: bool = False,
-    sum_over_batch: bool = False,
-    sum_over_timesteps: bool = True,
-) -> torch.Tensor:
-    r"""Average or sum over the respective dimensions of :attr:`sequence`, which
-    is of shape `[batch_size, max_time, ...]`.
-
-    Assumes :attr:`sequence` has been properly masked according to
-    :attr:`sequence_length`.
-
-    Args:
-        sequence: A tensor to reduce.
-        sequence_length: A tensor of shape `[batch_size]`. Time steps beyond
-            the respective sequence lengths will be made zero. If `None`,
-            no masking is performed.
-        average_across_batch (bool): If set, average the sequence across the
-            batch dimension. Must not set `average_across_batch`'
-            and `sum_over_batch` at the same time.
-        average_across_timesteps (bool): If set, average the sequence across
-            the time dimension. Must not set `average_across_timesteps`
-            and `sum_over_timesteps` at the same time.
-        sum_over_batch (bool): If set, sum the sequence across the
-            batch dimension. Must not set `average_across_batch`
-            and `sum_over_batch` at the same time.
-        sum_over_timesteps (bool): If set, sum the sequence across the
-            time dimension. Must not set `average_across_timesteps`
-            and `sum_over_timesteps` at the same time.
-
-    Returns:
-        A tensor with dimension reduction.
-    """
-    if average_across_timesteps and sum_over_timesteps:
-        raise ValueError("Only one of `average_across_timesteps` and " "`sum_over_timesteps` can be set.")
-    if average_across_batch and sum_over_batch:
-        raise ValueError("Only one of `average_across_batch` and " "`sum_over_batch` can be set.")
-
-    if sum_over_timesteps:
-        sequence = torch.sum(sequence, dim=1)
-    elif average_across_timesteps:
-        sequence = torch.sum(sequence, dim=1).float() / sequence_length.float()
-
-    if sum_over_batch:
-        sequence = torch.sum(sequence, dim=0)
-    elif average_across_batch:
-        sequence = torch.mean(sequence, dim=0)
-
-    return sequence
-
-
-def get_lengths_mask(X: FloatTensor, lengths: LongTensor) -> FloatTensor:
-
-    if any(
-        [X.ndim != 2, lengths.ndim != 1, X.shape[0] != lengths.shape[0]],
-    ):
-        raise ValueError
-
-    return sequence_mask(lengths, max_len=X.shape[1])
-
-
-def sequence_mask(
-    lengths: Union[torch.LongTensor, List[int]],
-    max_len: Optional[int] = None,
-    dtype: Optional[torch.dtype] = None,
-    device: Optional[torch.device] = None,
-) -> torch.ByteTensor:
-    r"""Return a mask tensor representing the first N positions of each cell.
-
-    If ``lengths`` has shape ``[d_1, d_2, ..., d_n]`` the resulting tensor
-    ``mask`` has dtype ``dtype`` and shape ``[d_1, d_2, ..., d_n, maxlen]``,
-    with
-
-    ```
-    mask[i_1, i_2, ..., i_n, j] = (j < lengths[i_1, i_2, ..., i_n])
-    ```
-
-    Examples:
-
-    ```python
-    sequence_mask([1, 3, 2], 5)  # [[True, False, False, False, False],
-                                 #  [True,  True,  True, False, False],
-                                 #  [True,  True, False, False, False]]
-
-    sequence_mask([[1, 3],[2,0]])  # [[[ True, False, False],
-                                   #   [ True,  True,  True]],
-                                   #  [[ True,  True, False],
-                                   #   [False, False, False]]]
-    ```
-
-    Args:
-        lengths: integer tensor or list of int, all its values <= max_len.
-        max_len: scalar integer tensor, size of last dimension of returned
-            tensor. Default is the maximum value in ``lengths``.
-        dtype: the desired data type of returned tensor. Default: if None,
-            returns :torch:`ByteTensor`.
-        device: the desired device of returned tensor. Default: if None, uses
-            the current device for the default tensor type.
-    Returns:
-        A mask tensor of shape :python:`lengths.shape + (max_len,)`, cast to
-        specified dtype.
-    Raises:
-        ValueError: if ``max_len`` is not a scalar.
-    """
-    if not isinstance(lengths, torch.Tensor):
-        lengths = torch.tensor(lengths, device=device)
-    elif device is None:
-        device = lengths.device
-    lengths: torch.LongTensor
-    if max_len is None:
-        max_len = torch.max(lengths).item()
-
-    size = lengths.size()
-    row_vector = (
-        torch.arange(max_len, device=device, dtype=lengths.dtype).view(*([1] * len(size)), -1).expand(*size, max_len)
-    )
-    mask = (row_vector < lengths.unsqueeze(-1)).to(device=device)
-    if dtype is not None:
-        mask = mask.to(dtype=dtype)
-
-    return mask
-
-
-def mask_sequences(
-    sequence: Union[torch.Tensor, List[int]],
-    sequence_length: Union[torch.LongTensor, List[int]],
-    dtype: Optional[torch.dtype] = None,
-) -> torch.Tensor:
-    r"""Masks out sequence entries that are beyond the respective sequence
-    lengths. Masks along the time dimension.
-
-    :attr:`sequence` and :attr:`sequence_length` can either be python
-    arrays or Tensors, respectively. If both are Python arrays (or None), the
-    return will be a Python array as well.
-
-    Args:
-        sequence: A Tensor or Python array of sequence values.
-            If ``time_major==False`` (default), this must be a Tensor of shape
-            ``[batch_size, max_time, ...]``. The batch and time dimension is
-            exchanged if ``time_major==True``.
-        sequence_length: A Tensor or python array of shape ``[batch_size]``.
-            Time steps beyond the respective sequence lengths will be
-            made zero.
-        dtype (dtype): Type of :attr:`sequence`. If `None`, infer from
-            :attr:`sequence` automatically.
-        time_major (bool): The shape format of the inputs. If `True`,
-            :attr:`sequence` must have shape
-            ``[max_time, batch_size, ...]``.
-            If `False` (default), :attr:`sequence` must have
-            shape ``[batch_size, max_time, ...]``.
-
-    Returns:
-        The masked sequence, i.e., a Tensor or python array of the same shape
-        as :attr:`sequence` but with masked-out entries (set to zero).
-
-        If both :attr:`sequence` and :attr:`sequence_length` are python
-        arrays, the returned value is a python array as well.
-    """
-    if not torch.is_tensor(sequence):
-        sequence = torch.tensor(sequence, dtype=dtype)
-    sequence: torch.Tensor
-
-    rank = sequence.dim()
-    if rank < 2:
-        raise ValueError("`sequence` must be 2D or higher order.")
-
-    max_time = sequence.size(1)
-    if dtype is None:
-        dtype = sequence.dtype
-    mask = sequence_mask(sequence_length, max_time, dtype=dtype)
-    mask = mask.view(*mask.size(), *([1] * (rank - 2)))
-    sequence = sequence * mask
-
-    return sequence
 
 
 def nested_detach_and_clone(obj: Any, to_cpu: bool = False, to_numpy: bool = False):
@@ -973,20 +750,18 @@ def nested_detach_and_clone(obj: Any, to_cpu: bool = False, to_numpy: bool = Fal
     return nested_tensor_operation(obj=obj, tensor_operation=_operation)
 
 
-def nested_to_cuda(obj: Any):
-    def _operation(X: torch.Tensor) -> torch.Tensor:
-        return X.cuda()
-
-    return nested_tensor_operation(obj=obj, tensor_operation=_operation)
-
-
-def nested_tensor_operation(obj: Any, tensor_operation: Callable[[torch.Tensor], Any]) -> Any:
+def nested_tensor_operation(
+    obj: Any, tensor_operation: Callable[[torch.Tensor], Any]
+) -> Any:
     """Nested Application of `detach().clone()`.
 
     This function will remove gradients and reference.
     """
     if isinstance(obj, (list, tuple)):
-        return [nested_tensor_operation(obj=_obj, tensor_operation=tensor_operation) for _obj in obj]
+        return [
+            nested_tensor_operation(obj=_obj, tensor_operation=tensor_operation)
+            for _obj in obj
+        ]
 
     if isinstance(obj, dict):
         new_dict_obj = {}
@@ -994,7 +769,9 @@ def nested_tensor_operation(obj: Any, tensor_operation: Callable[[torch.Tensor],
             if not isinstance(key, str):
                 raise NotImplementedError
 
-            new_dict_obj[key] = nested_tensor_operation(obj=val, tensor_operation=tensor_operation)
+            new_dict_obj[key] = nested_tensor_operation(
+                obj=val, tensor_operation=tensor_operation
+            )
 
         return new_dict_obj
